+++ tests/myria3d/models/test_model.py
@@ -29,11 +29,61 @@ def test_model_get_batch_tensor_by_enumeration():
         _ = model._get_batch_tensor_by_enumeration(batch.idx_in_original_cloud)
 
 
+
+def test_model_external_embeddings_get_batch_tensor_by_enumeration():
+    config = make_default_hydra_cfg(
+        overrides=[
+            "predict.src_las=tests/data/toy_dataset_src/862000_6652000.classified_toy_dataset.100mx100m.las",
+            "datamodule.epsg=2154",
+            "work_dir=./../../..",
+            "datamodule.subtile_width=1",  # Extreme case with very few points per subtile
+            "datamodule.hdf5_file_path=null",
+        ]
+    )
+
+    datamodule: LightningDataModule = hydra.utils.instantiate(config.datamodule)
+    datamodule._set_predict_data(config.predict.src_las)
+
+    model = Model(
+        neural_net_class_name="PyGRandLANet",
+        neural_net_hparams=dict(num_features=2, num_classes=7),
+        dims=[48, 192, 384, 768],
+    )
+    for batch in datamodule.predict_dataloader():
+        # Check that no error is raised ("TypeError: object of type 'numpy.int64' has no len()")
+        _ = model._get_batch_tensor_by_enumeration(batch.idx_in_original_cloud)
+
+
 def test_model_forward():
     config = make_default_hydra_cfg(
         overrides=[
             "predict.src_las=tests/data/toy_dataset_src/862000_6652000.classified_toy_dataset.100mx100m.las",
             "datamodule.epsg=2154",
+            "work_dir=./../../..",
+            "datamodule.subtile_width=1",  # Extreme case with very few points per subtile
+            "datamodule.hdf5_file_path=null",
+        ]
+    )
+
+    datamodule: LightningDataModule = hydra.utils.instantiate(config.datamodule)
+    datamodule._set_predict_data(config.predict.src_las)
+
+    model: LightningModule = hydra.utils.instantiate(config.model)
+    device = utils.define_device_from_config_param(config.predict.gpus)
+    model.to(device)
+    model.eval()
+    print(model.model)
+    for batch in datamodule.predict_dataloader():
+        # Check that no error is raised
+        targets, logits = model.forward(batch)
+
+
+def test_model_externalized_embeddings_forward():
+    config = make_default_hydra_cfg(
+        overrides=[
+            "predict.src_las=tests/data/toy_dataset_src/862000_6652000.classified_toy_dataset.100mx100m.las",
+            "datamodule.epsg=2154",
+            "model.neural_net_hparams.dims=[48,192,384,768]",
             "work_dir=./../../..",
             "datamodule.subtile_width=1",  # Extreme case with very few points per subtile
             "datamodule.hdf5_file_path=null",